requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Flask-Login==0.6.3
psycopg[binary]>=3.2.10
python-dotenv==1.0.1
gunicorn==22.0.0
